Fix USB buffer description table corruption
CCID_ENDP0_RXADDR was set to 0x18, so incoming buffers were written from
0x18 to 0x57 in the packet memory area (PMA).

This area contained the USB buffer description table (BTABLE) entry for
endpoint 4 (0x20 to 0x5F), defining its buffer address to 0x19C.

Data sent to the host with endpoint 4 was not the buffer located at the
defined address (0x19C) but some other data from the PMA, depending on
how the BTABLE was corrupted.

This prevented the functions defined in src/keyboard to work properly.